Move mutt_rename_file to lib/file.[hc]
