+++ rfc1524.h
@@ -50,6 +50,5 @@ void rfc1524_free_entry(struct Rfc1524MailcapEntry **entry);
 int rfc1524_expand_command(struct Body *a, char *filename, char *_type, char *command, int clen);
 int rfc1524_expand_filename(char *nametemplate, char *oldfile, char *newfile, size_t nflen);
 int rfc1524_mailcap_lookup(struct Body *a, char *type, struct Rfc1524MailcapEntry *entry, int opt);
-int mutt_rename_file(char *oldfile, char *newfile);
 
 #endif /* _MUTT_RFC1524_H */
